move types to .model file
