+++ libs/features/quest/src/quest-preview/quest-preview.component.ts
@@ -2,7 +2,8 @@ import { AsyncPipe } from '@angular/common';
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, DestroyRef, inject, OnInit } from '@angular/core';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { IconComponent } from '@keira/shared/base-editor-components';
-import { FactionRequirement, ItemObjective, NpcOrGoObjective, QuestObjectivesComponent } from './quest-objectives.component';
+import { QuestObjectivesComponent } from './quest-objectives.component';
+import { FactionRequirement, ItemObjective, NpcOrGoObjective } from './quest-objectives.model';
 import { RacesTextKey, RacesTextValue } from '@keira/shared/constants';
 import { Model3DViewerComponent, VIEWER_TYPE } from '@keira/shared/model-3d-viewer';
 import { PreviewHelperService } from '@keira/shared/preview';
